added serialization tests to test suite
